pyproject.toml
[project]
name = "sync-whoop-gsheets"
version = "0.1.0"
description = "CLI to sync WHOOP activities to Google Sheets"
authors = [
    { name = "Robin Weston" }
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "whoop",
    "gspread",
    "google-auth",
    "python-dotenv",
    "click",
    "flask",
    "requests-oauthlib"
]

[project.scripts]
sync-whoop-gsheets = "sync_whoop_to_gsheets:main"

[tool.uv] 